+++ src/components/index.ts
@@ -11,3 +11,12 @@ export type { FormAlertProps } from "./FormAlert";
 
 export { default as CookieNotice } from "./CookieNotice";
 export type { CookieNoticeProps } from "./CookieNotice";
+
+export { default as Spinner } from "./Spinner";
+export type { SpinnerProps } from "./Spinner";
+
+export { default as Skeleton } from "./Skeleton";
+export type { SkeletonProps } from "./Skeleton";
+
+export { default as SkeletonText } from "./SkeletonText";
+export type { SkeletonTextProps } from "./SkeletonText";